renaming retrying to reattempt
